+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [1.2.5] - 2019-09-17
 ### Fixed
  - Fix regression in 1.2.3 that broke detection of custom Proton installations.
  - Proton prefix is detected correctly even if it exists in a different Steam library folder than the game installation.
@@ -72,7 +72,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
  - The last release of protontricks maintained by [@Sirmentio](https://github.com/Sirmentio).
 
-[Unreleased]: https://github.com/Matoking/protontricks/compare/1.2.4...HEAD
+[Unreleased]: https://github.com/Matoking/protontricks/compare/1.2.5...HEAD
+[1.2.5]: https://github.com/Matoking/protontricks/compare/1.2.4...1.2.5
 [1.2.4]: https://github.com/Matoking/protontricks/compare/1.2.3...1.2.4
 [1.2.3]: https://github.com/Matoking/protontricks/compare/1.2.2...1.2.3
 [1.2.2]: https://github.com/Matoking/protontricks/compare/1.2.1...1.2.2